rgw: split RGWDataAccess from rgw_tools.cc
that class is not rados specific, while the rest of the code
in rgw_tools.cc is
